# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hybrid firewall in Go using public blocklists",
        "title": "Meds: net healing",
        "contact": {
            "name": "cnaize",
            "url": "https://github.com/cnaize"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "v1.2.0"
    },
    "paths": {
...
        "/v1/ja3/blocklist/exclude": {
            "get": {
                "description": "get all excluded ja3 hashes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Get excluded ja3 hashes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetJA3Resp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert ja3 hashes to excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Upsert excluded ja3 hashes",
                "parameters": [
                    {
                        "description": "hashes to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertJA3Req"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove ja3 hashes from excludelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Remove excluded ja3 hashes",
                "parameters": [
                    {
                        "description": "hashes to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveJA3Req"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/ja3/blocklist/exclude/{hash}": {
            "get": {
                "description": "check if a ja3 hash is excluded",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Check excluded ja3 hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hash to check",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckJA3Resp"
                        }
                    }
                }
            }
        },
        "/v1/ja3/blocklist/include": {
            "get": {
                "description": "get all included ja3 hashes",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Get included ja3 hashes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.GetJA3Resp"
                        }
                    }
                }
            },
            "post": {
                "description": "upsert ja3 hashes to includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Upsert included ja3 hashes",
                "parameters": [
                    {
                        "description": "hashes to add",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UpsertJA3Req"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "remove ja3 hashes from includelist",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Remove included ja3 hashes",
                "parameters": [
                    {
                        "description": "hashes to remove",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RemoveJA3Req"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/ja3/blocklist/include/{hash}": {
            "get": {
                "description": "check if a ja3 hash is included",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "JA3"
                ],
                "summary": "Check included ja3 hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "hash to check",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.CheckJA3Resp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.CheckJA3Resp": {
            "type": "object",
            "properties": {
                "found": {
                    "type": "boolean"
                }
            }
        },
...
        "api.GetJA3Resp": {
            "type": "object",
            "properties": {
                "hashes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a50a861119aceb0ccc74902e8fddb618",
                        "534ce2dbc413c68e908363b5df0ae5e0"
                    ]
                }
            }
        },
...
        "api.RemoveJA3Req": {
            "type": "object",
            "properties": {
                "hashes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a50a861119aceb0ccc74902e8fddb618",
                        "534ce2dbc413c68e908363b5df0ae5e0"
                    ]
                }
            }
        },
...
        "api.UpsertJA3Req": {
            "type": "object",
            "properties": {
                "hashes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "a50a861119aceb0ccc74902e8fddb618",
                        "534ce2dbc413c68e908363b5df0ae5e0"
                    ]
                }
            }
        },
...
    }
}